tehtävät 4.15-4.22

+++ controllers/blogs.js
@@ -2,6 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+// helper function
+
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1})
   response.json(blogs)
@@ -9,7 +11,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.userId)
+  const user =  request.user
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -17,7 +19,6 @@ blogsRouter.post('/', async (request, response) => {
     likes: body.likes || 0,
     user: user._id
   })
-
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
@@ -25,6 +26,11 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blogToBeDeletedId = request.params.id
+  const userBlogs = request.user.blogs.map(blog => blog.toString())
+  if (!userBlogs.includes(blogToBeDeletedId)) {
+    return response.status(401).json({ error: 'User ids do not match.' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
